split up tests for cq master
