+++ .github/workflows/docker_publish.yml
@@ -38,7 +38,17 @@ jobs:
           restore-keys: |
             ${{ runner.os }}-single-buildx
 
-      - name: Build and push dependencies Docker image
+      - name: Build and push dependencies Docker image CQ=master
+        uses: docker/build-push-action@v2
+        with:
+          push: true
+          target: dependencies
+          tags: ghcr.io/fusion-energy/paramak_cq_master:dependencies
+          build-args: |
+            cq_version=master
+            paramak_version=${{ github.ref_name }}
+
+      - name: Build and push dependencies Docker image CQ=2.1
         uses: docker/build-push-action@v2
         with:
           push: true
